Added Find_Duplicates Script
This script will read the Json file & sift through the file. Product_Key is made up of (name, price) & we two arrays that are filled as new & duplicate thereby storing any duplicate into the duplicates file. BUT it does not remove the duplicate from original file only shows which are duplicates.